+++ app/src/main/java/com/nextcloud/talk/chat/viewmodels/MessageInputViewModel.kt
@@ -45,13 +45,11 @@ class MessageInputViewModel @Inject constructor(
     }
 
     lateinit var chatRepository: ChatMessageRepository
-    lateinit var chatDataSource: ChatNetworkDataSource
     lateinit var currentLifeCycleFlag: LifeCycleFlag
     val disposableSet = mutableSetOf<Disposable>()
 
-    fun setData(chatMessageRepository: ChatMessageRepository, chatNetworkDataSource: ChatNetworkDataSource) {
+    fun setData(chatMessageRepository: ChatMessageRepository) {
         chatRepository = chatMessageRepository
-        chatDataSource = chatNetworkDataSource
     }
 
     override fun onResume(owner: LifecycleOwner) {
@@ -121,10 +119,6 @@ class MessageInputViewModel @Inject constructor(
     private val _callStartedFlow: MutableLiveData<Pair<ChatMessage, Boolean>> = MutableLiveData()
     val callStartedFlow: LiveData<Pair<ChatMessage, Boolean>>
         get() = _callStartedFlow
-
-    private val _maintenanceMode: MutableLiveData<Boolean> = MutableLiveData(false)
-    val maintenanceMode: LiveData<Boolean>
-        get() = _maintenanceMode
 
     @Suppress("LongParameterList")
     fun sendChatMessage(
@@ -208,18 +202,6 @@ class MessageInputViewModel @Inject constructor(
                 message,
                 editedMessageText
             ).collect {}
-        }
-    }
-
-    fun checkMaintenance(baseUrl: String) {
-        val url = baseUrl + ApiUtils.getUrlPostfixForStatus()
-        viewModelScope.launch {
-            try {
-                val status = chatDataSource.getServerStatus(url)
-                _maintenanceMode.postValue(status.maintenance)
-            } catch (_: Exception) {
-                _maintenanceMode.postValue(false)
-            }
         }
     }
