+++ app/src/main/res/layout/fragment_message_input.xml
@@ -25,17 +25,6 @@
         android:visibility="gone"
         tools:visibility="visible" />
 
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/fragment_maintenance_mode"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:text="@string/nc_dialog_maintenance_mode_description"
-        android:textColor="@color/white"
-        android:background="@color/hwSecurityRed"
-        android:visibility="gone"
-        tools:visibility="visible" />
-
     <include
         android:id="@+id/fragment_call_started"
         layout="@layout/call_started_message"
