+++ app/src/main/res/values/strings.xml
@@ -667,6 +667,7 @@ How to translate with transifex:
     <string name="nc_dialog_outdated_client_option_update">Update</string>
     <string name="nc_switch_account">Switch account</string>
     <string name="nc_dialog_maintenance_mode_description">Server is currently in maintenance mode.</string>
+    <string name="nc_dialog_maintenance_mode">Server is in maintenance mode. Please try again later.</string>
 
     <!-- Take photo -->
     <string name="take_photo">Take a photo</string>
